+++ tailcfg/tailcfg_test.go
@@ -5,13 +5,11 @@ package tailcfg_test
 
 import (
 	"encoding/json"
-	"log"
 	"net/netip"
 	"os"
 	"reflect"
 	"regexp"
 	"strconv"
-	"strings"
 	"testing"
 	"time"
 
@@ -795,57 +793,6 @@ func TestNodeIsRouter(t *testing.T) {
 			}
 		})
 	}
-}
-
-func FuzzNodeIsRouter(f *testing.F) {
-	encodePrefixes := func(f *testing.F, prefixes ...netip.Prefix) string {
-		f.Helper()
-		out := make([]string, len(prefixes))
-		for i, p := range prefixes {
-			out[i] = p.String()
-		}
-		return strings.Join(out, " ")
-	}
-	decodePrefixes := func(t *testing.T, prefixes string) []netip.Prefix {
-		t.Helper()
-		var out []netip.Prefix
-		for p := range strings.FieldsSeq(prefixes) {
-			pfx, err := netip.ParsePrefix(p)
-			if err != nil {
-				log.Printf("skipping %q: %v", prefixes, err)
-				t.Skipf("%q: %v", prefixes, err)
-			}
-			out = append(out, pfx)
-		}
-		return out
-	}
-
-	for _, tc := range nodeIsRouterCases {
-		addresses := encodePrefixes(f, tc.node.Addresses...)
-		allowedIPs := encodePrefixes(f, tc.node.AllowedIPs...)
-		f.Logf("addresses=%q allowedIPs=%q", addresses, allowedIPs)
-		f.Add(addresses, allowedIPs)
-	}
-	f.Fuzz(func(t *testing.T, addresses, allowedIPs string) {
-		n := Node{
-			Addresses:  decodePrefixes(t, addresses),
-			AllowedIPs: decodePrefixes(t, allowedIPs),
-		}
-		ps := peerStatusFromNode(n.View())
-		t.Logf("%v %v", n.Addresses, n.AllowedIPs)
-
-		if len(n.Addresses) != len(ps.TailscaleIPs) ||
-			len(n.AllowedIPs) != ps.AllowedIPs.Len() {
-			t.Skip("n and ps are not equivalent")
-		}
-
-		gotN := n.IsRouter()
-		gotPS := ps.IsRouter()
-		if gotN != gotPS {
-			t.Errorf("mismatched node %t, peer status %t; addresses=%q allowedIPs=%q",
-				gotN, gotPS, addresses, allowedIPs)
-		}
-	})
 }
 
 func peerStatusFromNode(n NodeView) *ipnstate.PeerStatus {
