+++ util/safediff/diff_test.go
@@ -178,19 +178,3 @@ func TestLines(t *testing.T) {
 		t.Errorf("Lines: output unexpectedly truncated")
 	}
 }
-
-func FuzzDiff(f *testing.F) {
-	f.Fuzz(func(t *testing.T, x, y string, maxSize int) {
-		const maxInput = 1e3
-		if len(x) > maxInput {
-			x = x[:maxInput]
-		}
-		if len(y) > maxInput {
-			y = y[:maxInput]
-		}
-		diff, _ := Lines(x, y, maxSize) // make sure this does not panic
-		if strings.Count(diff, "\n") > 1 && maxSize >= 0 && len(diff) > maxSize {
-			t.Fatal("maxSize exceeded")
-		}
-	})
-}
